+++ src/content-script/hooks/useFetchCollections.ts
@@ -1,5 +1,6 @@
 import { useQuery, UseQueryOptions } from "@tanstack/react-query";
 
+import usePplxAuth from "@/content-script/hooks/usePplxAuth";
 import { useGlobalStore } from "@/content-script/session-store/global";
 import PplxApi from "@/services/PplxApi";
 import { Collection } from "@/types/collection.types";
@@ -7,7 +8,7 @@ import { Collection } from "@/types/collection.types";
 export default function useFetchCollections({
   ...props
 }: Omit<UseQueryOptions<Collection[]>, "queryKey" | "queryFn"> = {}) {
-  const isLoggedIn = useGlobalStore((state) => state.isLoggedIn);
+  const { isLoggedIn } = usePplxAuth();
   const isReady = useGlobalStore((state) => state.isWebSocketCaptured);
 
   return useQuery<Collection[]>({