+++ src/content-script/session-store/global.ts
@@ -2,12 +2,10 @@ import { create } from "zustand";
 
 type GlobalState = {
   isWebSocketCaptured: boolean;
-  isLoggedIn: boolean;
 };
 
 const useGlobalStore = create<GlobalState>(() => ({
   isWebSocketCaptured: false,
-  isLoggedIn: false,
 }));
 
 const globalStore = useGlobalStore;