+++ src/cplx-user-settings/components/FloatingTrigger.tsx
@@ -3,6 +3,7 @@ import { HTMLAttributes } from "react";
 import useRouter from "@/content-script/hooks/useRouter";
 import Cplx from "@/shared/components/icons/Cplx";
 import Tooltip from "@/shared/components/Tooltip";
+import useCplxUserSettings from "@/shared/hooks/useCplxUserSettings";
 import { cn } from "@/utils/cn";
 import { whereAmI } from "@/utils/utils";
 
@@ -11,22 +12,43 @@ export default function FloatingTrigger({
 }: HTMLAttributes<HTMLDivElement>) {
   const location = whereAmI(useRouter().url);
 
+  const { data, mutation: updateCplxUserSettings } = useCplxUserSettings();
+
+  const isFirstVisit = data?.data?.isFirstVisit;
+
+  const handleInformed = () => {
+    if (!isFirstVisit) return;
+
+    updateCplxUserSettings.mutate((draft) => {
+      draft.isFirstVisit = false;
+    });
+  };
+
   return (
     <div
       className={cn(
         "tw-fixed tw-right-4 tw-z-0 tw-font-sans tw-transition-all tw-animate-in tw-fade-in",
         {
           "tw-bottom-[4rem] tw-hidden lg:tw-block": location === "thread",
-          "tw-bottom-[5rem] md:tw-bottom-[4rem] tw-block": location !== "thread",
+          "tw-bottom-[5rem] tw-block md:tw-bottom-[4rem]":
+            location !== "thread",
         },
       )}
+      onMouseEnter={handleInformed}
+      onClick={handleInformed}
       {...props}
     >
       <Tooltip
-        content="Settings"
+        content={
+          isFirstVisit
+            ? "Click to open the Settings menu for Complexity!"
+            : "Settings"
+        }
         positioning={{
           placement: "left",
         }}
+        portal={false}
+        defaultOpen={isFirstVisit}
       >
         <div
           className={cn(